+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## X.Y.Z (YYYY-MM-DD)
+
+  * Add methods for getting jobs from workers.
+  * Add method for getting jobs from the queue.
+  * Add method for creating a job from a hash.
+
 ## 0.1.0 (2020-08-10)
 
   * Change initialize.